Properly support self-addressed lamps.
In sent commands, the bottom bit of the address byte determines whether this
is a DAPC (Direct Arc Power Control) command or not. If the bit is zero, it's
a DAPC command, otherwise it's not.

We were previously dealing with this by, in reAddressLamps, setting the bottom
bit and in sendDapc, clearing it. This worked fine if using reAddressLamps, but
was broken if someone was addressing their own lamps.

Switch to not setting the bit in reAddressLamps, setting it when sending all
other commands and not doing anything in sendDapc.